+++ src/components/calculator.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import * as calc from '../utils/calculations-tmp';
+import * as calc from '../utils/calculations';
 
 /**
  * A class that returns my calculator project
