+++ src/index.js
@@ -1,7 +1,7 @@
 import './styles/style.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app-tmp';
+import App from './components/app';
 
 ReactDOM.render(<App />, document.getElementById('root'));
